Create the FirstCarousel container
Transfer all the necessary code from the App.js to FirstCarousel Container.
